KAS-5033 added route to strip signatures for 1 piece
