+++ config/piece.js
@@ -6,7 +6,9 @@ import {
   query,
   update,
 } from 'mu';
-import { APPLICATION_GRAPH } from '../cfg';
+import { APPLICATION_GRAPH, KANSELARIJ_GRAPH, FILE_RESOURCE_BASE } from '../cfg';
+import removeSignatures from '../lib/remove-signatures';
+import { createMuFile, pathToShareUri, readMuFile, writeMuFile } from '../lib/file';
 
 const PIECE_RESOURCE_BASE = process.env.PIECE_RESOURCE_BASE || 'http://themis.vlaanderen.be/id/stuk/';
 const ACCESS_LEVEL_PUBLIC = 'http://themis.vlaanderen.be/id/concept/toegangsniveau/c3de9c70-391e-4031-a85e-4b03433d6266';
@@ -146,8 +148,67 @@ WHERE {
   await updateFunction(queryString);
 }
 
+async function stripSignaturesFromPiece(pieceUri, graph=KANSELARIJ_GRAPH, queryFunction=query, updateFunction=update) {
+  if (!await isMainPiece(pieceUri, graph, queryFunction)) {
+    throw new Error(`Quad with subject <${pieceUri}> is not a "main" piece and we will not treat it further`);
+  }
+
+  const file = await getFileFromPiece(pieceUri, graph, queryFunction);
+  if (file === null || (file?.format.indexOf('application/pdf') === -1 && file?.extension?.toLowerCase() !== 'pdf')) {
+    throw new Error(`Quad with subject <${pieceUri}> does not have a file or the file isn't a PDF, not processing further`);
+  }
+
+  const pdfBytes = readMuFile(file);
+  const pdfBytesWithoutSignatures = await removeSignatures(pdfBytes);
+
+  if (pdfBytesWithoutSignatures === null) {
+    throw new Error(`Quad with subject <${pieceUri}> did not have signatures in its file, not processing further`);
+  }
+
+  console.log(`Generated PDF without signatures for signed file ${file.uri}`);
+
+  const { id: physicalFileUuid, path: pdfWithoutSignaturesPath} = writeMuFile(pdfBytesWithoutSignatures);
+  console.log(`Wrote PDF without signatures to ${pdfWithoutSignaturesPath}`);
+
+  const now = new Date();
+  const virtualFileUuid = uuid();
+  const virtualFile = {
+    id: virtualFileUuid,
+    uri: FILE_RESOURCE_BASE + virtualFileUuid,
+    name: `${file.pieceName}.pdf`,
+    extension: 'pdf',
+    size: pdfBytesWithoutSignatures.byteLength,
+    created: now,
+    format: 'application/pdf',
+  };
+
+  const physicalFile = {
+    id: physicalFileUuid,
+    uri: pathToShareUri(pdfWithoutSignaturesPath),
+    name: `${physicalFileUuid}.pdf`,
+    extension: 'pdf',
+    size: pdfBytesWithoutSignatures.byteLength,
+    created: now,
+    format: 'application/pdf',
+  };
+
+  console.log(`Creating virtual mu-file ${virtualFile.uri}`);
+  await createMuFile(virtualFile, physicalFile, graph, updateFunction);
+
+  console.log(`Linking virtual mu-file ${virtualFile.uri} to piece ${pieceUri}`);
+  await linkSignatureStrippedPDFToPiece(
+    pieceUri,
+    file.uri,
+    virtualFile.uri,
+    !!file.derivedFile,
+    graph,
+    updateFunction
+  );
+}
+
 export {
   isMainPiece,
   getFileFromPiece,
   linkSignatureStrippedPDFToPiece,
+  stripSignaturesFromPiece,
 }
